+++ src/main/scala/model/core/SimulationState.scala
@@ -3,7 +3,7 @@ package model.core
 import model.cure.Cure
 import model.infection.InfectionAndDeathPopulation.*
 import model.plague.Disease
-import model.plague.db.Symptoms.pulmonaryEdema
+import model.plague.traits.Symptoms.pulmonaryEdema
 import model.reaction.Reactions
 import model.reaction.Reactions.StandardReactions
 import model.time.TimeTypes.{Day, Year}
