+++ src/test/scala/plague/DiseaseTest.scala
@@ -1,7 +1,7 @@
 package plague
 import model.plague.{Disease, Trait}
-import model.plague.db.Symptoms.*
-import model.plague.db.Symptoms
+import model.plague.traits.Symptoms.*
+import model.plague.traits.Symptoms
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
